Moved fermi and bose functions to lattice_utility.cpp
