+++ c++/triqs_tprf/lattice/lattice_utility.hpp
@@ -39,4 +39,24 @@ namespace triqs_tprf {
   @return Tuple of chi_dyn_wk, the dynamic part of chi :math:`\chi^{(dyn)}_{abcd}(i\omega_n, \mathbf{k})`, which converges to zero for :math:`\omega_n \rightarrow \infty`, and chi_const_k, the part of chi that is constant in Matsubara frequency space :math:`\chi^{(stat)}_{abcd}(\mathbf{k})`.
   */
   std::tuple<chi_wk_t, chi_k_t> split_into_dynamic_wk_and_constant_k(chi_wk_cvt chi_wk);
+
+  /** Helper function to evaluate the Fermi-Dirac distribution function
+
+  .. math ::
+      f(\epsilon) = \frac{1}{\exp(\epsilon) + 1}
+  
+  @param e : point at which to evaluate :math:`f(\epsilon)`.
+  @return The value of :math:`f(\epsilon)`.
+  */
+  double fermi(double e);
+
+  /** Helper function to evaluate the Bose-Einstein distribution function
+
+  .. math ::
+      n_B(\epsilon) = \frac{1}{\exp(\epsilon) - 1}
+  
+  @param e : point at which to evaluate :math:`n_B(\epsilon)`.
+  @return The value of :math:`n_B(\epsilon)`.
+  */
+  double bose(double e);
 } // namespace triqs_tprf
